Step2: Move pilot theme and console islands into place
The theme files (theme, RThemeManager, RThemeBackend, and the three
per-platform backends) and the console files (the console and terminal
dock widgets, history, and the syntax highlighter and rules) each form
a closed island with no scene or drawable edges, so they move wholesale
into theme/ and console/.

Beyond the include-path rewrite and the pilot CMake list regrouping,
the gtest list references three of the moved sources directly for the
Qt-free suites; those paths update in gtests/CMakeLists.txt.  No logic
changes.
