+++ cpp/solvcon/pilot/console/RPythonConsoleDockWidget.hpp
@@ -15,8 +15,8 @@
 
 #include <solvcon/python/common.hpp> // must be first.
 
-#include <solvcon/pilot/RPythonConsoleHistory.hpp>
-#include <solvcon/pilot/RPythonSyntaxHighlighter.hpp>
+#include <solvcon/pilot/console/RPythonConsoleHistory.hpp>
+#include <solvcon/pilot/console/RPythonSyntaxHighlighter.hpp>
 
 #include <cstddef>
 #include <string>
