+++ cpp/solvcon/pilot/console/RPythonConsoleHistory.hpp
