+++ cpp/solvcon/pilot/console/RPythonSyntaxHighlighter.cpp
@@ -3,9 +3,9 @@
  * BSD 3-Clause License, see COPYING
  */
 
-#include <solvcon/pilot/RPythonSyntaxHighlighter.hpp>
+#include <solvcon/pilot/console/RPythonSyntaxHighlighter.hpp>
 
-#include <solvcon/pilot/RPythonSyntaxRules.hpp>
+#include <solvcon/pilot/console/RPythonSyntaxRules.hpp>
 
 #include <QColor>
 #include <QTextBlock>
