+++ cpp/solvcon/pilot/console/RPythonSyntaxHighlighter.hpp
@@ -13,7 +13,7 @@
  * @ingroup group_domain
  */
 
-#include <solvcon/pilot/theme.hpp>
+#include <solvcon/pilot/theme/theme.hpp>
 
 #include <functional>
