+++ cpp/solvcon/pilot/console/RPythonSyntaxRules.hpp
