+++ gtests/CMakeLists.txt
@@ -40,9 +40,9 @@ add_executable(
     test_nopython_pilot_syntax.cpp
     test_nopython_pilot_theme.cpp
     test_nopython_pilot_keymap.cpp
-    ${SOLVCON_INCLUDE_DIR}/solvcon/pilot/RPythonConsoleHistory.cpp
-    ${SOLVCON_INCLUDE_DIR}/solvcon/pilot/RPythonSyntaxRules.cpp
-    ${SOLVCON_INCLUDE_DIR}/solvcon/pilot/theme.cpp
+    ${SOLVCON_INCLUDE_DIR}/solvcon/pilot/console/RPythonConsoleHistory.cpp
+    ${SOLVCON_INCLUDE_DIR}/solvcon/pilot/console/RPythonSyntaxRules.cpp
+    ${SOLVCON_INCLUDE_DIR}/solvcon/pilot/theme/theme.cpp
     ${SOLVCON_INCLUDE_DIR}/solvcon/pilot/keymap.cpp
     ${SOLVCON_TOGGLE_SOURCES}
     ${SOLVCON_PROFILING_SOURCES}
